fix: moved methods to utils
