+++ india_compliance/gst_india/utils/e_invoice.py
@@ -319,6 +319,18 @@ def validate_if_e_invoice_can_be_cancelled(doc):
         )
 
 
+def get_e_invoice_info(doc):
+    e_invoice_info = frappe.db.get_value(
+        "e-Invoice Log", doc.irn, ("invoice_data", "acknowledged_on"), as_dict=True
+    )
+    invoice_data = e_invoice_info.pop("invoice_data")
+    if not invoice_data:
+        return e_invoice_info, None
+    invoice_data = frappe.parse_json(invoice_data)
+    seller_gstin = invoice_data.SellerDtls.get("Gstin")
+    return e_invoice_info, seller_gstin
+
+
 class EInvoiceData(GSTTransactionData):
     def get_data(self):
         self.validate_transaction()
